[#3231] project_editor: Extract utility functions into new module
